refactor: update import paths for toast to use hooks directory
