+++ src/components/features/admin/addons/AddonsTable.tsx
@@ -6,7 +6,7 @@ import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { DataTable } from '@/components/tables/addonChecks/data-table';
 import { Button } from '@/components/ui/button.tsx';
 import { useFetchAddons, useUpdateAddon } from '@/api/appwrite/useAddons';
-import { toast } from '@/api';
+import { toast } from '@/hooks';
 import type { Addon } from '@/types';
 import { Switch } from '@/components/ui/switch';
